+++ changelog.md
@@ -6,8 +6,7 @@
 This version is not released yet and is under active development.
 ```
 
-- Replace `pylint`, `pycln` and `pydocstyle` with `ruff`.
-- Run `pyupgrade` twice.
+- Replace `pyupgrade`, `pylint`, `pycln` and `pydocstyle` with `ruff`.
 
 ## [2.7.6 (2023-02-13)](https://github.com/kdeldycke/workflows/compare/v2.7.5...v2.7.6)